+++ Kohi/Models/OutboundModel.cs
@@ -11,7 +11,7 @@ namespace Kohi.Models
     {
         public int Id { get; set; }              // Mã phiếu xuất
         public int InventoryId { get; set; }     // Mã lô hàng (liên kết với InventoryModel)
-        public int Quantity { get; set; }        // Số lượng xuất từ lô này
+        public float Quantity { get; set; }        // Số lượng xuất từ lô này
         public DateTime OutboundDate { get; set; } // Ngày xuất kho
         public string? Purpose { get; set; }      // Mục đích xuất (ví dụ: sản xuất, hư hỏng)
         public string? Notes { get; set; }        // Ghi chú (tùy chọn)
